# core-banking-psdx-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-banking-psdx
  description: PSD2/PSD3 and FIDA Regulatory Compliance Service
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/psdx
    description: Development Environment
  - url: /
    description: Local Development Environment
security:
  - apiKey: []
  - bearer: []
...
paths:
  /api/v1/providers/{providerId}:
    get:
      tags:
        - Third Party Provider Management
      summary: Get provider
      description: |
        ## Retrieve a Third Party Provider

        This endpoint retrieves detailed information about a specific Third Party Provider (TPP) by its unique identifier.

        ### Description
        This operation allows authorized users to access the complete profile of a registered TPP.

        ### Path Parameters
        * `providerId` - The unique identifier of the TPP to retrieve

        ### Processing
        The system will:
        1. Validate the provided TPP ID
        2. Check if the TPP exists in the system
        3. Verify that the requester has permission to view the TPP information

        ### Response
        Returns the complete TPP profile including:
        * Basic information (name, ID, status)
        * Contact details
        * Regulatory information
        * Registration date
        * Current status (active, suspended, etc.)

        ### Security Considerations
        Access to this endpoint should be restricted to authorized personnel only.
      operationId: getProvider
      parameters:
        - name: providerId
          in: path
          description: ID of the provider
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Request-ID
          in: header
          description: Unique ID for the request
          schema:
            type: string
          example: 99391c7e-ad88-49ec-a2ad-99ddcb1f7721
        - name: X-Consent-ID
          in: header
          description: ID of the consent for the operation
          schema:
            type: string
          example: '12345'
      responses:
        '200':
          description: Provider found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDThirdPartyProviderDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDThirdPartyProviderDTO'
        '404':
          description: Provider not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDThirdPartyProviderDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDThirdPartyProviderDTO'
    put:
      tags:
        - Third Party Provider Management
      summary: Update provider
      description: |
        ## Update a Third Party Provider

        This endpoint allows for updating the information of an existing Third Party Provider (TPP).

        ### Description
        This operation enables authorized users to modify the details of a registered TPP.
        It's used when TPP information changes or needs correction.

        ### Path Parameters
        * `providerId` - The unique identifier of the TPP to update

        ### Request Body
        The request must include the updated TPP information:
        * TPP name and identification details
        * Contact information
        * Regulatory information
        * Any other fields that need updating

        ### Processing
        The system will:
        1. Validate the provided TPP ID
        2. Check if the TPP exists in the system
        3. Verify that the requester has permission to update the TPP
        4. Validate the updated information
        5. Apply the changes to the TPP record

        ### Response
        Returns the updated TPP profile with all changes applied.

        ### Validation Rules
        * The TPP must exist in the system
        * Required fields cannot be empty
        * Regulatory information must be valid

        ### Security Considerations
        * Only authorized administrators should be able to update TPP information
        * All changes are logged for audit purposes
      operationId: updateProvider
      parameters:
        - name: providerId
          in: path
          description: ID of the provider
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Request-ID
          in: header
          description: Unique ID for the request
          schema:
            type: string
          example: 99391c7e-ad88-49ec-a2ad-99ddcb1f7721
        - name: X-Consent-ID
          in: header
          description: ID of the consent for the operation
          schema:
            type: string
          example: '12345'
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PSDThirdPartyProviderDTO'
        required: true
      responses:
        '200':
          description: Provider updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDThirdPartyProviderDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDThirdPartyProviderDTO'
        '404':
          description: Provider not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDThirdPartyProviderDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDThirdPartyProviderDTO'
    delete:
      tags:
        - Third Party Provider Management
      summary: Revoke provider
      description: |
        ## Revoke a Third Party Provider

        This endpoint permanently revokes a Third Party Provider's (TPP) access to the system.

        ### Description
        This operation allows authorized administrators to permanently disable a TPP's ability
        to access the APIs. This is typically used when a TPP is no longer authorized to provide
        services, has violated terms of service, or has ceased operations.

        ### Path Parameters
        * `providerId` - The unique identifier of the TPP to revoke

        ### Processing
        The system will:
        1. Validate the provided TPP ID
        2. Check if the TPP exists in the system
        3. Verify that the requester has permission to revoke TPPs
        4. Permanently revoke the TPP's access
        5. Invalidate all tokens and credentials
        6. Mark the TPP record as "REVOKED" in the database

        ### Response
        Returns a boolean value indicating successful revocation (true).

        ### Business Rules
        * A revoked TPP cannot access any APIs
        * All existing tokens and credentials for the TPP are permanently invalidated
        * Revocation is a permanent action and cannot be easily reversed
        * A revoked TPP would need to re-register to regain access

        ### Security Considerations
        * Only authorized administrators should be able to revoke TPPs
        * The action is logged for audit purposes
        * Notification should be sent to the TPP and relevant regulatory authorities
        * All customer consents associated with the TPP should be reviewed
      operationId: revokeProvider
      parameters:
        - name: providerId
          in: path
          description: ID of the provider
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Request-ID
          in: header
          description: Unique ID for the request
          schema:
            type: string
          example: 99391c7e-ad88-49ec-a2ad-99ddcb1f7721
        - name: X-Consent-ID
          in: header
          description: ID of the consent for the operation
          schema:
            type: string
          example: '12345'
      responses:
        '200':
          description: Provider revoked successfully
          content:
            application/json:
              schema:
                type: boolean
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: boolean
        '404':
          description: Provider not found
          content:
            application/json:
              schema:
                type: boolean
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: boolean
...
  /api/v1/providers/{providerId}/suspend:
    post:
      tags:
        - Third Party Provider Management
      summary: Suspend provider
      description: |
        ## Suspend a Third Party Provider

        This endpoint temporarily suspends a Third Party Provider's (TPP) access to the system.

        ### Description
        This operation allows authorized administrators to temporarily disable a TPP's ability
        to access the APIs. This is typically used when there are security concerns, regulatory
        issues, or at the request of the TPP itself.

        ### Path Parameters
        * `providerId` - The unique identifier of the TPP to suspend

        ### Processing
        The system will:
        1. Validate the provided TPP ID
        2. Check if the TPP exists in the system
        3. Verify that the requester has permission to suspend TPPs
        4. Change the TPP's status to "SUSPENDED"
        5. Invalidate any active sessions for the TPP

        ### Response
        Returns the updated TPP profile with status changed to "SUSPENDED".

        ### Business Rules
        * A suspended TPP cannot access any APIs
        * All existing tokens for the TPP are invalidated
        * The suspension can be reversed using the activate endpoint

        ### Security Considerations
        * Only authorized administrators should be able to suspend TPPs
        * The action is logged for audit purposes
        * Notification may be sent to the TPP and relevant regulatory authorities
      operationId: suspendProvider
      parameters:
        - name: providerId
          in: path
          description: ID of the provider
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Request-ID
          in: header
          description: Unique ID for the request
          schema:
            type: string
          example: 99391c7e-ad88-49ec-a2ad-99ddcb1f7721
        - name: X-Consent-ID
          in: header
          description: ID of the consent for the operation
          schema:
            type: string
          example: '12345'
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Provider suspended successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDThirdPartyProviderDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDThirdPartyProviderDTO'
        '404':
          description: Provider not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDThirdPartyProviderDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDThirdPartyProviderDTO'
...
components:
  schemas:
...
    PSDThirdPartyProviderDTO:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the TPP
          format: uuid
        name:
          type: string
          description: Name of the TPP
          example: FinTech Solutions Ltd
        registrationNumber:
          type: string
          description: Registration number of the TPP
          example: TPP123456
        nationalCompetentAuthority:
          type: string
          description: National competent authority ID
          example: DE-BAFIN
        nationalCompetentAuthorityCountry:
          type: string
          description: National competent authority ID
          example: DE
        redirectUri:
          type: string
          description: Redirect URI for the TPP
          example: https://fintech-solutions.com/callback
        status:
          type: string
          description: Status of the TPP
          example: ACTIVE
        providerType:
          type: string
          description: Type of the TPP
          example: AISP
        roles:
          type: array
          description: Roles of the TPP
          items:
            type: string
            description: Roles of the TPP
        certificate:
          $ref: '#/components/schemas/PSDCertificateDTO'
        get_links:
          $ref: '#/components/schemas/PSDLinksDTO'
        createdAt:
          type: string
          description: Date and time when the TPP was created
          format: date-time
        updatedAt:
          type: string
          description: Date and time when the TPP was last updated
          format: date-time
      description: PSD Third Party Provider information
...
  securitySchemes:
    apiKey:
      type: apiKey
      description: API key for TPP authentication
      name: X-API-KEY
      in: header
    bearer:
      type: http
      description: JWT token for user authentication
      scheme: bearer
      bearerFormat: JWT